add ToHighlightedString to simplify conversions
