+++ internal/search/commit/commit_test.go
@@ -1,11 +1,8 @@
 package commit
 
 import (
-	"strconv"
 	"testing"
 	"testing/quick"
-
-	"github.com/stretchr/testify/require"
 
 	"github.com/sourcegraph/sourcegraph/internal/search/result"
 )
@@ -43,71 +40,5 @@ func TestCommitSearchResult_Limit(t *testing.T) {
 				t.Error("small exhaustive check failed")
 			}
 		}
-	}
-}
-
-func Test_searchRangeToHighlights(t *testing.T) {
-	type testCase struct {
-		input      string
-		inputRange result.Range
-		output     []result.HighlightedRange
-	}
-
-	cases := []testCase{
-		{
-			input: "abc",
-			inputRange: result.Range{
-				Start: result.Location{Offset: 1, Line: 0, Column: 1},
-				End:   result.Location{Offset: 2, Line: 0, Column: 2},
-			},
-			output: []result.HighlightedRange{{
-				Line:      0,
-				Character: 1,
-				Length:    1,
-			}},
-		},
-		{
-			input: "abc\ndef\n",
-			inputRange: result.Range{
-				Start: result.Location{Offset: 2, Line: 0, Column: 2},
-				End:   result.Location{Offset: 5, Line: 1, Column: 1},
-			},
-			output: []result.HighlightedRange{{
-				Line:      0,
-				Character: 2,
-				Length:    1,
-			}, {
-				Line:      1,
-				Character: 0,
-				Length:    1,
-			}},
-		},
-		{
-			input: "abc\ndef\nghi\n",
-			inputRange: result.Range{
-				Start: result.Location{Offset: 0, Line: 0, Column: 0},
-				End:   result.Location{Offset: 11, Line: 2, Column: 3},
-			},
-			output: []result.HighlightedRange{{
-				Line:      0,
-				Character: 0,
-				Length:    3,
-			}, {
-				Line:      1,
-				Character: 0,
-				Length:    3,
-			}, {
-				Line:      2,
-				Character: 0,
-				Length:    3,
-			}},
-		},
-	}
-
-	for i, tc := range cases {
-		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			res := searchRangeToHighlights(tc.input, tc.inputRange)
-			require.Equal(t, tc.output, res)
-		})
 	}
 }
